If the input had a value on load, don't remove it on tabbing through the field
One possible side-effect of the [DHainzl-Seekda/typeahead.js@3fceb14](https://github.com/DHainzl-Seekda/typeahead.js/commit/3fceb14d01cf606652ed10008c8f648d8b110386) fix may happen if you are using a form for editing a record and auto populate the field with the record's value. Once the user opens the form, the value is shown on the field. Once "tabbing" through the field, it would get erased by this fix.

Here we enhance previous fix by, on tab keyed, checking if the input originally had a value and, if so, select it.

_DISCLAIMER_
This fix __doesn't__ cover all the cases. It works if the original field value is present  on the autocomplete menu. Tested on a `local` dataset with `limit` superior to the total number of possible entries. If the original value is not present on the menu, it would still get erased. e.g the original "State" value is "Texas" and the `limit` is 5 (default)